Next and Previous button now are not hidden anymore
They get enabled and disabled in base of which slice of gifs you are exploring.

Created DisablePreviousButtonsChecker and DisableNextButtonsChecker
This two functions return true of falose in base of which slice are you exploring. They results
are then used to enable or disable the buttons.

Add a p tag which points the number of the page. Repositioned Buttons and page pointer with CSS inline